+++ packages/learn-card-base/CHANGELOG.md
@@ -1,5 +1,19 @@
 # learn-card-base
 
+## 0.1.11
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/lca-api-plugin@1.1.8
+
+## 0.1.10
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/lca-api-plugin@1.1.7
+
 ## 0.1.9
 
 ### Patch Changes